requirements.txt
google-cloud-texttospeech
pydub
google-generativeai
requests
sounddevice
numpy
scipy
websockets
asyncio
pyinstaller
